# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.CreateLecturerRequest:
    properties:
      department:
        type: string
      lecturer_id:
        type: string
      user_id:
        type: string
    required:
    - lecturer_id
    - user_id
    type: object
  model.CreateLecturerResponse:
    properties:
      data:
        $ref: '#/definitions/model.Lecturer'
      status:
        type: string
    type: object
  model.CreateStudentRequest:
    properties:
      academic_year:
        type: string
      advisor_id:
        type: string
      program_study:
        type: string
      student_id:
        type: string
      user_id:
        type: string
    required:
    - student_id
    - user_id
    type: object
  model.CreateStudentResponse:
    properties:
      data:
        $ref: '#/definitions/model.Student'
      status:
        type: string
    type: object
  model.CreateUserRequest:
    properties:
      academic_year:
        type: string
      advisor_id:
        type: string
      department:
        type: string
      email:
        type: string
      full_name:
        type: string
      is_active:
        type: boolean
      lecturer_id:
        type: string
      password:
        type: string
      program_study:
        type: string
      role_id:
        type: string
      student_id:
        type: string
      username:
        type: string
    required:
    - email
    - full_name
    - password
    - role_id
    - username
    type: object
  model.CreateUserResponse:
    properties:
      data:
        $ref: '#/definitions/model.User'
      status:
        type: string
    type: object
...
  model.DeleteUserResponse:
    properties:
      status:
        type: string
    type: object
...
  model.GetAllRolesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Role'
        type: array
      status:
        type: string
    type: object
...
  model.GetAllUsersResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.User'
        type: array
      status:
        type: string
    type: object
...
  model.GetUserByIDResponse:
    properties:
      data:
        $ref: '#/definitions/model.User'
      status:
        type: string
    type: object
...
  model.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  model.UpdateUserRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      is_active:
        type: boolean
      role_id:
        type: string
      username:
        type: string
    required:
    - email
    - full_name
    - role_id
    - username
    type: object
  model.UpdateUserResponse:
    properties:
      data:
        $ref: '#/definitions/model.User'
      status:
        type: string
    type: object
  model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      password_hash:
        type: string
      role_id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:3001
info:
  contact: {}
  description: API untuk sistem pelaporan prestasi mahasiswa dengan dukungan multi-role
    (Admin, Dosen Wali, Mahasiswa) dan field prestasi yang fleksibel
  title: Sistem Pelaporan Prestasi Mahasiswa API
  version: "1.0"
paths:
...
  /roles:
    get:
      consumes:
      - application/json
      description: Mengambil daftar semua role dari database. Hanya dapat diakses
        oleh admin dengan permission user:manage
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetAllRolesResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get all roles
      tags:
      - Users
...
  /users:
    get:
      consumes:
      - application/json
      description: Mengambil daftar semua user dari database. Hanya dapat diakses
        oleh admin dengan permission user:manage
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetAllUsersResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Membuat user baru di database. Dapat sekaligus membuat student
        atau lecturer profile. Hanya dapat diakses oleh admin dengan permission user:manage
      parameters:
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CreateUserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create user
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus user berdasarkan ID. Hanya dapat diakses oleh admin dengan
        permission user:manage
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DeleteUserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Mengambil data user spesifik berdasarkan ID. Hanya dapat diakses
        oleh admin dengan permission user:manage
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetUserByIDResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Memperbarui data user berdasarkan ID. Hanya dapat diakses oleh
        admin dengan permission user:manage
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UpdateUserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update user
      tags:
      - Users
  /users/{id}/lecturer-profile:
    post:
      consumes:
      - application/json
      description: Membuat lecturer profile untuk user tertentu. User harus memiliki
        role Dosen Wali. Hanya dapat diakses oleh admin dengan permission user:manage
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Lecturer data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateLecturerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CreateLecturerResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create lecturer profile
      tags:
      - Users
  /users/{id}/role:
    put:
      consumes:
      - application/json
      description: Memperbarui role user berdasarkan ID. Hanya dapat diakses oleh
        admin dengan permission user:manage
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Role ID
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update user role
      tags:
      - Users
  /users/{id}/student-profile:
    post:
      consumes:
      - application/json
      description: Membuat student profile untuk user tertentu. User harus memiliki
        role Mahasiswa. Hanya dapat diakses oleh admin dengan permission user:manage
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Student data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateStudentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CreateStudentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create student profile
      tags:
      - Users
schemes:
- http
securityDefinitions:
  Bearer:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"